MORE

+++ utils/afni.cpp
@@ -821,7 +821,7 @@ MRI *afniRead(const char *fname, int read_volume)
           {
             pmem = (void *)malloc(bytes_per_voxel * mri->width);
             if (pmem) {
-              if (fread(pmem, bytes_per_voxel, mri->width, fp) != mri->width) {
+              if ((int)fread(pmem, bytes_per_voxel, mri->width, fp) != mri->width) {
                 fclose(fp);
                 MRIfree(&header);
                 errno = 0;
